Change file names to switch
