+++ global_buffer/Makefile
@@ -33,7 +33,8 @@ GLB_TOP_SDF ?= gls/glb.sdf
 GLB_TILE_SDF ?= gls/glb_tile.sdf
 
 # runtime
-RUN_ARGS ?= +test1 +test2 +test3 +test4 +test5 +test6 +test7 +test8 +test9
+# RUN_ARGS ?= +test1 +test2 +test3 +test4 +test5 +test6 +test7 +test8 +test9
+RUN_ARGS ?= +test1
 RUN_ARGS += +MAX_NUM_TEST=9
 RUN_ARGS += +DEBUG
 RUN_LOG ?= run.log
