+++ global_buffer/sim/glb_test.sv
@@ -9,9 +9,7 @@ program glb_test (
     // LEFT
     input  logic                      clk,
     input  logic                      reset,
-    output logic [ NUM_GLB_TILES-1:0] core_stall,
-    output logic [ NUM_GLB_TILES-1:0] rtr_stall,
-    output logic [ NUM_GLB_TILES-1:0] pcfg_stall,
+    output logic [ NUM_GLB_TILES-1:0] pcfg_broadcast_stall,
     output logic [NUM_CGRA_TILES-1:0] cgra_stall_in,
 
     // proc
@@ -130,7 +128,7 @@ program glb_test (
                 end
             end
         end
-        glb_stall(test.core_stall_mask, test.rtr_stall_mask, test.pcfg_stall_mask);
+        glb_pcfg_broadcast_stall(test.pcfg_broadcast_stall_mask);
 
         foreach (kernels[i]) begin
             if (kernels[i].type_ == WR) begin
@@ -342,9 +340,7 @@ program glb_test (
 
     task initialize();
         // control
-        core_stall <= 0;
-        rtr_stall <= 0;
-        pcfg_stall <= 0;
+        pcfg_broadcast_stall <= 0;
         cgra_stall_in <= 0;
         pcfg_start_pulse <= 0;
         strm_g2f_start_pulse <= 0;
@@ -387,16 +383,10 @@ program glb_test (
         repeat (10) @(posedge clk);
     endtask
 
-    task glb_stall(logic [NUM_GLB_TILES-1:0] core_tile_mask,
-                   logic [NUM_GLB_TILES-1:0] rtr_tile_mask,
-                   logic [NUM_GLB_TILES-1:0] pcfg_tile_mask);
+    task glb_pcfg_broadcast_stall(logic [NUM_GLB_TILES-1:0] pcfg_broadcast_stall_mask);
 
-        #2 $display("Glb tiles CORE logics are stalled with mask %16b", core_tile_mask);
-        core_stall <= core_tile_mask;
-        $display("Glb tiles ROUTER logics are stalled with mask %16b", rtr_tile_mask);
-        rtr_stall <= rtr_tile_mask;
-        $display("Glb tiles PCFGROUTER logics are stalled with mask %16b", pcfg_tile_mask);
-        pcfg_stall <= pcfg_tile_mask;
+        #2 $display("Glb tiles PCFG broadcast logics are stalled with mask %16b", pcfg_broadcast_stall_mask);
+        pcfg_broadcast_stall <= pcfg_broadcast_stall_mask;
         repeat (4) @(posedge clk);
 
     endtask
